Create ShowDetails Component
